Add y_range_min_size theme parameter
Automatic finding of y ranges results in ts always
being scaled to fill as much of the plot height as possible.
If this is undesired (e.g. when trying to accentuate that a
series is relatively stable) this parameter allows fixing the
total range of y values to a certain minimal value.