Add check for length 1 time series
These cause all sorts of problems which are not obviously connected to
the fact that length 1 series are inputted. Plotting single points does
not make much sense anyway so it just stops if any series of length 1
are provided.

Maybe it would be possible to support sets of time series of which only
some are length 1 but is that necessary?

Fixes #134