Correct subtitle-device-size-drift
An offset that is based on par("mex") is added in the C-code for
whatever reason. Since mex is set by tsplot based on device size, this
offset will change accordingly. It is now being subtracted in add_title.

Note: the facor of 0.2 is technically device dependent but is 0.2 for
most of them.