Allow adding an extra tick exactly at 0
The "is adding an extra tick ok or does it cross 0 when it shouldn't"
test uses sign() to determine zero crossings.
However, adding a tick AT zero would be deemed ok but the test still
fails since sign(0) == 0. This takes that into account.

Kinda treating the sympton rather than the caust though.

Closes #223