Fix x axis being overdrawn with two axis plot
The right lower bound was not adjusted in tsplot when adding extra ticks
due to the left axis needing some. Then the old upper/lower bound of the
right axis would be used to check if it also needs extra ticks,
resulting in the same tick being added twice but addYGrids only omits
the first ELEMENT of the ticks to be drawn, not the first VALUE -> a
grid line was drawn over the x axis.

Fixes #227