make long_to_ts handle series of length 1
Previously it would crash when such series appeared in the
data. Straight out stripping them might not always be the
desired behavior.

Another issue with the na.trim step here (in addition to the
numeric imprecisions it introduces in the dates) is that it
appears to round up the year of the (single observation) series.